refactor(actions): move actions into semantic modules
