+++ lua/nvim-tree/actions/init.lua
@@ -342,15 +342,15 @@ local DEFAULT_MAPPING_CONFIG = {
 }
 
 function M.setup(opts)
-  require("nvim-tree.actions.system-open").setup(opts)
-  require("nvim-tree.actions.trash").setup(opts)
-  require("nvim-tree.actions.open-file").setup(opts)
-  require("nvim-tree.actions.change-dir").setup(opts)
-  require("nvim-tree.actions.create-file").setup(opts)
-  require("nvim-tree.actions.rename-file").setup(opts)
-  require("nvim-tree.actions.remove-file").setup(opts)
-  require("nvim-tree.actions.copy-paste").setup(opts)
-  require("nvim-tree.actions.expand-all").setup(opts)
+  require("nvim-tree.actions.fs.trash").setup(opts)
+  require("nvim-tree.actions.node.system-open").setup(opts)
+  require("nvim-tree.actions.node.open-file").setup(opts)
+  require("nvim-tree.actions.root.change-dir").setup(opts)
+  require("nvim-tree.actions.fs.create-file").setup(opts)
+  require("nvim-tree.actions.fs.rename-file").setup(opts)
+  require("nvim-tree.actions.fs.remove-file").setup(opts)
+  require("nvim-tree.actions.fs.copy-paste").setup(opts)
+  require("nvim-tree.actions.tree-modifiers.expand-all").setup(opts)
 
   cleanup_existing_mappings()
   M.mappings = vim.deepcopy(DEFAULT_MAPPINGS)
