+++ lua/nvim-tree/actions/tree-modifiers/collapse-all.lua
