+++ lua/nvim-tree/lib.lua
@@ -69,7 +69,7 @@ end
 
 local function handle_buf_cwd(cwd)
   if M.respect_buf_cwd and cwd ~= core.get_cwd() then
-    require("nvim-tree.actions.change-dir").fn(cwd)
+    require("nvim-tree.actions.root.change-dir").fn(cwd)
   end
 end
 
@@ -107,18 +107,18 @@ function M.open(cwd)
   view.restore_tab_state()
 end
 
--- @deprecated: use nvim-tree.actions.collapse-all.fn
-M.collapse_all = require("nvim-tree.actions.collapse-all").fn
--- @deprecated: use nvim-tree.actions.dir-up.fn
-M.dir_up = require("nvim-tree.actions.dir-up").fn
--- @deprecated: use nvim-tree.actions.change-dir.fn
-M.change_dir = require("nvim-tree.actions.change-dir").fn
--- @deprecated: use nvim-tree.actions.reloaders.reload_explorer
-M.refresh_tree = require("nvim-tree.actions.reloaders").reload_explorer
--- @deprecated: use nvim-tree.actions.reloaders.reload_git
-M.reload_git = require("nvim-tree.actions.reloaders").reload_git
--- @deprecated: use nvim-tree.actions.find-file.fn
-M.set_index_and_redraw = require("nvim-tree.actions.find-file").fn
+-- @deprecated: use nvim-tree.actions.tree-modifiers.collapse-all.fn
+M.collapse_all = require("nvim-tree.actions.tree-modifiers.collapse-all").fn
+-- @deprecated: use nvim-tree.actions.root.dir-up.fn
+M.dir_up = require("nvim-tree.actions.root.dir-up").fn
+-- @deprecated: use nvim-tree.actions.root.change-dir.fn
+M.change_dir = require("nvim-tree.actions.root.change-dir").fn
+-- @deprecated: use nvim-tree.actions.reloaders.reloaders.reload_explorer
+M.refresh_tree = require("nvim-tree.actions.reloaders.reloaders").reload_explorer
+-- @deprecated: use nvim-tree.actions.reloaders.reloaders.reload_git
+M.reload_git = require("nvim-tree.actions.reloaders.reloaders").reload_git
+-- @deprecated: use nvim-tree.actions.finders.find-file.fn
+M.set_index_and_redraw = require("nvim-tree.actions.finders.find-file").fn
 
 function M.setup(opts)
   M.hijack_unnamed_buffer_when_opening = opts.hijack_unnamed_buffer_when_opening
